Dispatch multi layers, but still not showing both
